refactor: reorganize project into src/, public/, styles/ structure
- Move cartogram.js and line-chart.js into src/
- Move data files (JSON, CSV) into public/data/ (served at runtime)
- Move us-atlas TopoJSON into public/ (renamed to us-atlas-10m.json)
- Move SVG icons into public/static/
- Remove now-empty d3-playground/ directory entirely
- Remove local topojson-client.min.js (installed from npm in TICKET-02)
- Update all import and fetch paths to match new layout
- Add mise.toml for Node.js version pinning
- Verify vite build copies data files to dist/
