+++ index.html
@@ -134,6 +134,6 @@
       </div>
     </div>
 
-    <script type="module" src="./d3-playground/modules/cartogram-chart/cartogram.js"></script>
+    <script type="module" src="./src/cartogram.js"></script>
   </body>
 </html>
